requirements.txt
Django==2.0
django-mdeditor==0.1.14
Markdown==3.1.1
pkg-resources==0.0.0
Pygments==2.4.2
pytz==2019.1
